Add support for reporting device NUMA node
In order to allow the pod scheduling to be topology-aware, we need to
report the NUMA node that each device is associated with.

The initial corresponding API change in Kubernetes is tracked under:
https://github.com/kubernetes/kubernetes/pull/74423 and there are
a couple of subsequent changes after that.

And now that we've got the ability to query the device NUMA node,
we can correctly set the existing "numa" field as well.